lattice
Module framework

Modules

animation

Animation module providing timing-synchronized animation and transition controls.

audio

⚠️ Experimental Receive single-channel, multiband audio with configurable FFT bands.

audio_controls

Control sketch parameters with audio signals. Supports any number of channels of the device you can specify in MULTICHANNEL_AUDIO_DEVICE_NAME.

automate

Supporting types for the Animation::automate method, providing a breakpoint animation system for creating complex automation curves, similar to those found in DAWs, yet with some special tricks of its own.

control_provider

control_script

Provides a means of controlling sketch parameters with the various Lattice control systems from an external yaml file that can be hot-reloaded.

controls

A generic abstraction over UI control structures

distance

easing

Easing functions, most of which borrowed from easings.net, which in turn come from Robert Penner, the guy who literally wrote the book on easings.

effects

Signal processing effects designed to operate on the results of Animation methods but may be suitable for other domains.

frame_controller

Provides a hand-rolled frame rate / counting singleton for syncing video recording, animations, and rendering with nannou. Nannou never implemented frame rate so here we are. The implementation is technically flawed but so far has been working well enough for my purposes (animations are tight and videos seem perfectly synced). The module is meant for internal framework/runtime use and should not be interacted with directly.

gpu

logging

midi

midi_controls

noise

osc_controls

osc_receiver

prelude

sketch

timing

A timing abstraction built to allow various syncing mechanisms for Lattice’s Animation system, including:

util

window_rect