lattice::framework
Module audio_controls

Control sketch parameters with audio signals. Supports any number of channels of the device you can specify in MULTICHANNEL_AUDIO_DEVICE_NAME.

Structs

AudioControlBuilder

AudioControlConfig

AudioControlState 🔒

AudioControls

MultichannelAudioProcessor 🔒

Functions

default_buffer_processor

The default processor used in AudioControls that applies a pre-emphasis filter to the incoming buffer then reduces that buffer into a single usable value via peak or RMS amplitude detection.

thru_buffer_processor

A “dummy” processor that simply returns that last value of the audio buffer, leaving it unprocessed. Use this for CV.

Type Aliases

BufferProcessor

A function used in AudioControls to reduce a channel’s audio buffer to a single value suitable for parameter control. The default_buffer_processor is specifically for audio-rate signals, while thru_buffer_processor should be used for control-rate audio signals (CV). You can also pass you own custom processor.