lattice::framework
Module automate

Supporting types for the Animation::automate method, providing a breakpoint animation system for creating complex automation curves, similar to those found in DAWs, yet with some special tricks of its own.

Structs

Breakpoint

The core structure need to configure segments for the Animation::automate method. See the various constructors such as Breakpoint::step, Breakpoint::ramp, etc. for in depth details.

Enums

Constrain

Kind

Mode

Shape