lattice::framework::control_provider
Trait ControlProvider

pub trait ControlProvider {
    // Required methods
    fn as_controls_mut(&mut self) -> &mut Controls;
    fn items(&self) -> &Vec<Control>;
    fn items_mut(&mut self) -> &mut Vec<Control>;
    fn update_value(&mut self, name: &str, value: ControlValue);
    fn to_serialized(&self) -> SerializedControls;
}

Shared trait to enable sketches to use either Controls or ControlScript instances via a single Model.controls struct field and have them automatically show up in the UI in either case.

Required Methods

fn as_controls_mut(&mut self) -> &mut Controls

fn items(&self) -> &Vec<Control>

fn items_mut(&mut self) -> &mut Vec<Control>

fn update_value(&mut self, name: &str, value: ControlValue)

fn to_serialized(&self) -> SerializedControls