lattice::framework::control_script
Module config

Deserialization types needed for converting the Lattice yaml format into controls

Structs

AudioConfig

AutomateConfig

BreakpointConfig

CheckboxConfig

ControlConfig

EffectConfig

MidiConfig

ModulationConfig

OscConfig

SelectConfig

Separator 🔒

Shared 🔒

SliderConfig

TriangleConfig

Enums

AnimationConfig

ControlType

EffectKind

KeyframeSequence

KindConfig

MaybeControlConfig

Functions

default_easing 🔒

default_false 🔒

default_iterations 🔒

default_mode 🔒

default_none_string 🔒

default_normalized_range 🔒

default_param_value_0 🔒

default_param_value_0_3 🔒

default_param_value_0_5 🔒

default_param_value_0_7 🔒

default_param_value_0_25 🔒

default_param_value_1 🔒

default_shape 🔒

deserialize_number_or_none 🔒

Type Aliases

ConfigFile

Uses IndexMap so we maintain the exact order of UI controls that are declared in yaml