lattice::framework::control_script
Module param_mod

Types and trait implementations to support parameter modulation.

Example

In the following example, a 4 beat triangle wave is used as the value of a wave_folder’s symmetry param.

t1:
  type: triangle
  beats: 4
  range: [-1, 1]

t2:
  type: wave_folder
  symmetry: $t1

See the parameter handling documentation for details on how different parameter types are processed. [link]: https://github.com/Lokua/lattice/blob/main/docs/parameter_handling.md

Enums

ParamValue

Traits

FromColdParams

Used for part 1 of an Effect’s instantiation phase

SetFromParam

Functions

warn_for 🔒