lattice::framework::control_script::param_mod
Trait FromColdParams

pub trait FromColdParams: Default + SetFromParam {
    // Provided method
    fn from_cold_params(config: &EffectConfig) -> Self { ... }
}

Used for part 1 of an Effect’s instantiation phase

Provided Methods

fn from_cold_params(config: &EffectConfig) -> Self

Extract the f32s from ParamValue::Cold variants and sets them on a newly created Effect instance. Will use the Effect’s default instead of ParamValue::Hot since those are swapped in during ControlScript::get. Important that this only deals with ParamValue (f32) - you still need to deal with copying the non-ParamValues from the EffectConfig to the Effect instance manually.

IMPORTANT!

This may suffer from the same issue that was addressed in this commit (non-param fields being stuck with their defaults instead of the config values)

Object Safety

This trait is not object safe.