lattice::framework
Module easing

Easing functions, most of which borrowed from easings.net, which in turn come from Robert Penner, the guy who literally wrote the book on easings.

Enums

Easing

Constants

C1 🔒

C2 🔒

C3 🔒

C4 🔒

C5 🔒

Functions

bounce_out 🔒

ease_in_back

ease_in_bounce

ease_in_circ

ease_in_cubic

ease_in_elastic

ease_in_expo

ease_in_out_back

ease_in_out_bounce

ease_in_out_circ

ease_in_out_cubic

ease_in_out_elastic

ease_in_out_expo

ease_in_out_quad

ease_in_out_quart

ease_in_out_quint

ease_in_out_sine

ease_in_quad

ease_in_quart

ease_in_quint

ease_in_sine

ease_out_back

ease_out_circ

ease_out_cubic

ease_out_elastic

ease_out_expo

ease_out_quad

ease_out_quart

ease_out_quint

ease_out_sine

exponential

linear

logarithmic

sigmoid

Suggested range [1, 10]