lattice::framework
Module frame_controller

Provides a hand-rolled frame rate / counting singleton for syncing video recording, animations, and rendering with nannou. Nannou never implemented frame rate so here we are. The implementation is technically flawed but so far has been working well enough for my purposes (animations are tight and videos seem perfectly synced). The module is meant for internal framework/runtime use and should not be interacted with directly.

Structs

FrameController 🔒

Statics

CONTROLLER 🔒

Functions

advance_single_frame

average_fps

clear_force_render

ensure_controller

fps

frame_count

is_paused

reset_frame_count

set_fps

set_frame_count

set_paused

wrapped_update

wrapped_view