lattice::framework::gpu
Struct GpuState

pub struct GpuState<V: Pod + Zeroable> {

    render_pipeline: RenderPipeline,
    vertex_buffer: Option<Buffer>,
    params_buffer: Buffer,
    params_bind_group: BindGroup,
    n_vertices: u32,
    depth_texture: Option<TextureView>,
    depth_stencil: Option<DepthStencilState>,
    topology: PrimitiveTopology,
    blend: Option<BlendState>,
    params_bind_group_layout: BindGroupLayout,
    vertex_buffers: Vec<VertexBufferLayout<'static>>,
    sample_count: u32,
    window_size_physical: [u32; 2],
    texture_bind_group: BindGroup,
    texture_bind_group_layout: BindGroupLayout,
    _marker: PhantomData<V>,
    update_state: Arc<Mutex<Option<PathBuf>>>,
    _watcher: Option<RecommendedWatcher>,

}

Fields
render_pipeline: RenderPipelinevertex_buffer: Option<Buffer>params_buffer: Bufferparams_bind_group: BindGroupn_vertices: u32depth_texture: Option<TextureView>depth_stencil: Option<DepthStencilState>topology: PrimitiveTopologyblend: Option<BlendState>params_bind_group_layout: BindGroupLayoutvertex_buffers: Vec<VertexBufferLayout<'static>>sample_count: u32window_size_physical: [u32; 2]texture_bind_group: BindGrouptexture_bind_group_layout: BindGroupLayout_marker: PhantomData<V>update_state: Arc<Mutex<Option<PathBuf>>>_watcher: Option<RecommendedWatcher>
Implementations

impl<V: Pod + Zeroable + Typed> GpuState<V>

pub fn new<P: Pod + Zeroable>( app: &App, window_size_logical: [u32; 2], shader_path: PathBuf, params: &P, vertices: Option<&[V]>, topology: PrimitiveTopology, blend: Option<BlendState>, enable_depth_testing: bool, watch: bool, ) -> Self

Creates a new GPU state manager with custom vertex data.

Type Parameters

V - The vertex type that defines the structure of vertex data sent to the GPU. Must derive Pod, Zeroable, and Reflect. While the implementation uses the Typed trait, this is automatically implemented when deriving Reflect.

window_size - The logical width and height of the target window. This will be scaled internally to match the primary display’s pixel depth.

See a full example at src/sketches/genuary_2025/g25_18_wind.rs. See the specialized new_procedural and new_full_screen constructors for easier to get up and running shaders.

fn create_params_bind_group_layout<P: Pod>(device: &Device) -> BindGroupLayout

fn create_params_buffer<P: Pod>(device: &Device, params: &P) -> Buffer

fn create_params_bind_group( device: &Device, layout: &BindGroupLayout, buffer: &Buffer, ) -> BindGroup

fn create_vertex_buffer(device: &Device, vertices: &[V]) -> Buffer

fn create_vertex_buffer_layout() -> VertexBufferLayout<'static>

fn create_texture_bind_group_layout(device: &Device) -> BindGroupLayout

fn create_render_pipeline(state: PipelineCreationState<'_>) -> RenderPipeline

fn start_shader_watcher( path: PathBuf, state: Arc<Mutex<Option<PathBuf>>>, ) -> RecommendedWatcher

pub fn set_input_texture(&mut self, app: &App, texture_view: &TextureView)

pub fn update<P: Pod>( &mut self, app: &App, window_size: [u32; 2], params: &P, vertices: &[V], )

For non-procedural and full-screen shaders when vertices are altered on CPU

pub fn update_params<P: Pod>( &mut self, app: &App, window_size: [u32; 2], params: &P, )

For procedural and full-screen shaders that do not need updated vertices

pub fn update_vertex_buffer( &mut self, app: &App, window_size: [u32; 2], vertices: &[V], )

fn update_shader(&mut self, app: &App)

Safely checks if the shader code has been modified then updates in it in place only if it has. If parsing or validation fails for any reason the method will return early and we will keeping using the last version

fn validate_shader(&self, shader_content: &str) -> Result<Module, ()>

fn recreate_pipeline(&mut self, app: &App, shader_content: &str)

fn check_and_handle_resize(&mut self, app: &App, window_size: [u32; 2])

pub fn render(&self, frame: &Frame<'_>)

pub fn render_to_texture(&self, app: &App) -> TextureView

fn infer_vertex_attributes() -> Vec<VertexAttribute>

impl GpuState<BasicPositionVertex>

pub fn new_fullscreen<P: Pod + Zeroable>( app: &App, window_size: [u32; 2], shader_path: PathBuf, params: &P, watch: bool, ) -> Self

Specialized impl for shaders that simply need every pixel. See interference and wave_fract for examples.

impl GpuState<()>

pub fn new_procedural<P: Pod + Zeroable>( app: &App, window_size: [u32; 2], shader_path: PathBuf, params: &P, watch: bool, ) -> Self

Specialized impl for purly procedural shaders (no vertices). See spiral.rs for an example.

pub fn render_procedural(&self, frame: &Frame<'_>, vertex_count: u32)