lattice::framework
Module prelude

Re-exports

pub use crate::framework::gpu;

pub use crate::framework::midi;

pub use crate::framework::animation::*;

pub use crate::framework::audio::*;

pub use crate::framework::audio_controls::*;

pub use crate::framework::automate::*;

pub use crate::framework::control_provider::*;

pub use crate::framework::control_script::*;

pub use crate::framework::controls::*;

pub use crate::framework::distance::*;

pub use crate::framework::easing::*;

pub use crate::framework::effects::*;

pub use crate::framework::logging::*;

pub use crate::framework::midi_controls::*;

pub use crate::framework::noise::*;

pub use crate::framework::osc_controls::*;

pub use crate::framework::sketch::*;

pub use crate::framework::timing::*;

pub use crate::framework::util::*;

pub use crate::framework::window_rect::*;

Macros

debug_once

debug_throttled

Logs a debug message at most once within a specified time interval.

loud_panic

Helper to make panic errors stand out more in logs

string_vec

ternary

warn_once

Derive Macros

SketchComponents