lattice::framework
Module timing

A timing abstraction built to allow various syncing mechanisms for Lattice’s Animation system, including:

Internal frame counting

External MIDI Clock with SPP support

External MIDI Time Code

External Hybrid mechanism utilizing both MIDI Clock and MTC for when SPP isn’t supported

External OSC for syncing specifically with Ableton Live via MaxForLive (preferred)

Manual timing for generating visualizations of animation sequences statically

The core abstraction is the Timing enum which wraps various TimingSource implementations. In all cases you must provide a bpm parameter as like the Animation module, time is provided in musical beats, e.g. 1.0 = 1 quarter note, 0.5 = 1 eight note, and so on. This provides the most intuitive way to write animations that are in sync with music.

Structs

FrameTiming

HybridTiming

ManualTiming

Allows sketches to visualize animations statically by manually providing what beat we’re on. This is especially useful for visualizing Breakpoint sequences

MidiSongTiming

OscTransportTiming

Uses the Open Sound Protocol to sync with Ableton Live via the L.OSCTransport.amxd MaxForLive device included in this project’s source code on GitHub. See the project README’s OSC section for more information.

Enums

Timing

Wrapper for all TimingSource implementations which allows run-time selection of a TimingSource via command line argument. Sketches can bypass the command line argument by using a TimingSource other than this directly.

Constants

CLOCK

CONTINUE

MTC_QUARTER_FRAME 🔒

PULSES_PER_QUARTER_NOTE 🔒

SONG_POSITION

START

STOP

TICKS_PER_QUARTER_NOTE 🔒

Traits

TimingSource